# squeeze out more csv perf