# old

```
         21.019,35 msec task-clock                #    9,306 CPUs utilized          
             3.993      context-switches          #    0,190 K/sec                  
                22      cpu-migrations            #    0,001 K/sec                  
         1.411.263      page-faults               #    0,067 M/sec                  
    74.282.315.584      cycles                    #    3,534 GHz                    
   104.099.611.130      instructions              #    1,40  insn per cycle         
    17.954.097.738      branches                  #  854,170 M/sec                  
       164.844.992      branch-misses             #    0,92% of all branches        

       2,258783482 seconds time elapsed

      18,279651000 seconds user
       2,735462000 seconds sys

```

# new

```
         20.794,88 msec task-clock                #    9,414 CPUs utilized          
             4.865      context-switches          #    0,234 K/sec                  
                35      cpu-migrations            #    0,002 K/sec                  
         1.411.261      page-faults               #    0,068 M/sec                  
    74.147.131.389      cycles                    #    3,566 GHz                    
   103.234.409.677      instructions              #    1,39  insn per cycle         
    17.678.999.439      branches                  #  850,161 M/sec                  
       134.160.088      branch-misses             #    0,76% of all branches        

       2,208917338 seconds time elapsed

      18,148388000 seconds user
       2,643275000 seconds sys
```